Stage 2 Part 1 - Adding a more complex security Bearer Authentication
Add a new dependency to the build gradle to enable this API to become a resource server.

Add to the application.yml more configurations to allow JWTs passed into the server to be
verified.

Add to the getAllCashCardsForUser method the use of the security context, which is set
once a user is authenticated. From this we can pull their sub and use that to only pull
back their data instead of everyone elses.

Update tables, entities and dtos to mirror the tutorial more closely to help following along.

Update unit tests.