Stage 2 Part 2 - Enable a custom annotation to get the sub from the JWT
Using the custom annotation, we can pull the sub from the verified JWT.
Once we have this, we can use it in the POST saveCashCard method to ensure that
the user can only create cash cards for themselves. Thus, removing the need to pass in any
user information in the request body.

Also use custom annotation in the getAllCashCardsForUser to ensure they only get their own cash cards back.