Stage 3 Continued - Add Google as oauth2 provider with user registration.
Add the client registration on the application.yml

Add a custom oauth2 user service and add it to the security config which allows
spring on request, redirect to googles login page, and on successful login capture
the returned user information and create a user account on our service for that user.

Next to come will be HTML pages for loging in, out and also accessing cashcards.
Also more quality of life such as global exception handlers and logging etc.