Stage 4 Part 1 - Adding the basic UI layouts
In this stage, we've moved away from returning an API json response
and instead we're returning a page rendered with Thymeleaf. The basic
header makes use of thymelead security extras to determine if the current
user is authenticated and conditionally displays a login / logout nav bar menu
item.

In part 2 we'll clean up the UI, and add more pages with CSS and more.