Stage 4 Part 2 - Finalise UI
In this commit I finalise the current implementation of this oauth2 project.
We have now got a working example of handling users through oauth2, blocking
resources from being returned if they don't belong to the current user
and event some basic UI which changes depending on the user being authenticated.